Moved Screen class to compose package
