+++ app/src/main/java/com/google/samples/apps/sunflower/compose/SunflowerApp.kt
@@ -30,7 +30,6 @@ import com.google.samples.apps.sunflower.R
 import com.google.samples.apps.sunflower.compose.gallery.GalleryScreen
 import com.google.samples.apps.sunflower.compose.home.HomeScreen
 import com.google.samples.apps.sunflower.compose.plantdetail.PlantDetailsScreen
-import com.google.samples.apps.sunflower.utilities.Screen
 
 @Composable
 fun SunflowerApp() {
